requirements.txt
boto3==1.9.253
pandas==1.0.3
PyYAML==3.13